backend/requirements.txt
fastapi
uvicorn
websockets
groq
python-dotenv
edge-tts